requirements.txt
playwright==1.40.0
python-dotenv==1.0.0
apscheduler==3.10.4
requests==2.31.0
